+++ Adj/Algebra.hs
@@ -3,3 +3,4 @@ module Adj.Algebra (module Exports) where
 import Adj.Algebra.Semigroupoid as Exports
 import Adj.Algebra.Category as Exports
 import Adj.Algebra.Functor as Exports
+import Adj.Algebra.Morphism as Exports