+++ Adj/Algebra/Morphism/Straight.hs
@@ -1,0 +1,3 @@
+module Adj.Algebra.Morphism.Straight where
+
+newtype Straigth morphism source target = Straight (morphism source target)